requirements.txt
aiohttp-socks==0.8.3
loguru==0.7.0
pycoingecko==3.1.0
Requests==2.31.0
starknet-py==0.18.1
tqdm==4.66.1
web3==6.8.0
fake-useragent==1.1.3
aiohttp_proxy==0.1.2
ccxt==4.0.42
pyTelegramBotAPI==4.12.0
openpyxl==3.1.2